@@ -11,11 +11,12 @@ public class MalosTarget {
 
     public static final int DEVICEINFO = 20012;
     public static final int IMU        = 20013;
-    public static final int GPIO       = 20049;
     public static final int HUMIDITY   = 20017;
-    public static final int UV         = 20013 + (4 * 4);
-    public static final int EVERLOOP   = 20013 + 8;
-    public static final int SERVO      = 20013 + 32;
+    public static final int EVERLOOP   = 20021;
+    public static final int PRESSURE   = 20025;
+    public static final int UV         = 20029;
+    public static final int SERVO      = 20045;
+    public static final int GPIO       = 20049;
 
     private int baseport;
     private String host;
